kubernetes/tools/average_image/requirements.txt
numpy==1.16.2
Pillow==6.0.0
